refactor: adjust payment.service import
